fix: update text for running tests in course page and submission page [percy]
The text for running tests in the course page and submission page has been updated to provide more clarity. Now, it differentiates between running tests for the current step and running tests for other stages. Additionally, it also provides instructions on where to view the live logs based on how the submission was made.