Update CourseIdeaCardComponent and CourseIdeaModel
The CourseIdeaCardComponent now includes an EmberTooltip to provide additional information about recently added ideas.

In the CourseIdeaModel, the condition for determining if an idea is newly created has been modified. Previously, an idea was considered newly created if it was created within the last 30 days or had less than 20 votes. Now, an idea is considered newly created if it was created within the last 30 days or has less than 100 votes.