feat: add redirect functionality after completing onboarding survey
The changes in the code include:
- Added support for redirecting to a specific page after completing the onboarding survey.
- Added a test to verify that the redirect functionality works correctly.
- Updated the WelcomeController to handle the redirect after survey completion.

The redirect functionality allows the user to specify a `?next` query parameter in the URL, which will be used to redirect the user to the specified page after they have completed the onboarding survey. If no `?next` parameter is provided, the user will be redirected to the catalog page.

The WelcomeController now handles storing and redirecting to the `?next` parameter. If a valid URL is provided, the user will be redirected to that page. If an invalid URL is provided, an error message will be logged and the user will be redirected to the catalog page as a fallback.