feat: update onboarding survey test to include filling in free form input
The onboarding survey test in the acceptance test file `onboarding-survey-test.js` has been updated to include filling in a free form input field. This is done by calling the new `fillInFreeFormInput` function in the `onboardingSurveyWizard` page object. The test now fills in the free form input with the value "Dummy" and asserts that the value is saved correctly.