feat: add login tooltip for anonymous users in StageListItemComponent
The changes in this commit add a login tooltip for anonymous users in the StageListItemComponent. When an anonymous user hovers over a stage item, a tooltip will appear with the text "Login via GitHub to view this stage". This provides a prompt for anonymous users to log in and access the stage content.