feat: add submittedAt field to Factory default values
The submittedAt field is added to the default values of the Factory. This field is set to the current date and time when a new instance of the Factory is created. This change allows for easier tracking and management of the submission time for community solutions.

Also, remove unnecessary eslint-disable-next-line comment from the test case.