GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Refactor session model and mirage config
 (refs/gitbutler/refactor-session-model-and-mirage-config-)
   branch head: 5676bca04b7ff475576ef75b61ed4720a6a019c1
 - Update pay controller (refs/gitbutler/update-pay-controller)
   branch head: 11b11b98591b040d6c06db4299152b99270d9fa7
   - tests/acceptance/pay-test.js
   - app/components/header/account-dropdown.js
   - app/models/session.js
   - mirage/config.js
   - app/controllers/pay.ts
   - app/routes/pay.js
   - app/components/pay-page/configure-checkout-session-modal.ts
   - app/controllers/pay.js
   - app/routes/pay.ts
   - app/templates/pay.hbs

Your previous branch was: refs/heads/add-tracks-controller-and-catalog-route

The sha for that commit was: b2790304e0debfd861dba5c30d1f1b8e222726fc

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch