feat: add support for pending product walkthrough feature suggestion in IntroductionController
The changes in this commit add support for displaying a pending product walkthrough feature suggestion in the IntroductionController. The `IntroductionController` now declares the `authenticator` service and uses it to access the `currentUser` property. If the `currentUser` has a pending product walkthrough feature suggestion, it will be displayed in the template. This allows users to see and interact with pending feature suggestions specific to their account.