feat: add bookmark functionality to code examples in the UI
The changes in this commit add the ability to bookmark code examples in the UI. Two new buttons are added to the code example component: one for bookmarking the code example and another for removing the bookmark. The bookmarked code examples will be displayed at the top of the list for easy access. This functionality is currently implemented as a temporary feature for demonstration purposes.