feat: display and copy tree SHA in submission details
This commit adds functionality to display and copy the tree SHA in the
submission details section of the submissions page in the course admin
interface. The tree SHA is now included in the header of the submission
details and users can click on a button to copy the Git tree SHA to the
clipboard. This change provides easier access to the tree SHA information
for users reviewing submissions.