Refactor Mirage config and handlers
This commit refactors the Mirage config and handlers files. It removes unused imports and rearranges the import statements. It
also adds the syncRepositoryStageLists utility function to the repositories handler file. Additionally, it imports the add
function from the date-fns library in the referral-activations handler file. These changes improve the organization and
maintainability of the Mirage codebase.