Remove build_production_bundle job from .github/workflows/test.yml
This commit removes the build_production_bundle job from the .github/workflows/test.yml file. This job was running on a specific Ubuntu
version and was only triggered when the branch was main. The steps in this job involved checking out the code, setting up Node.js, running
npm ci to install dependencies, and building the project. The environment variables CODECOV_TOKEN and SENTRY_AUTH_TOKEN were also set during
the build process.