feat: update discount logic and improve test scenarios
Refactor the `activeDiscountForYearlyPlan` getter to specify the return 
type as `PromotionalDiscountModel | null`. Update acceptance tests to 
reflect changes in discount types from early bird to signup. Replace 
direct user updates with a sign-in method for better clarity and 
consistency. Add a wait for functionality in the `UpgradePromptComponent` 
to handle asynchronous discount fetching.