feat(pay-page): add pricingPlanId to checkout session model
Add pricingPlanId attribute to IndividualCheckoutSessionModel and update
ChooseMembershipPlanModal to include it when creating a new checkout session.
This change is done to replace the usage of pricingFrequency with an explicit
pricingPlanId, enabling more accurate tracking and processing of the selected
membership plan during checkout. The pricingFrequency field remains temporarily
for backward compatibility.