feat: add PATCH endpoint and editable message in gifts UI
Add a PATCH /membership-gifts/:id endpoint to update membership gift
attributes, enabling partial updates on existing gifts. Return a 404
error if the gift does not exist.

Make the gift message editable in the UI by adding a
data-test-gift-message-editable attribute to the message container.
This improves testing and allows inline editing of the gift message.