feat(gifts): add "Buy another gift" button to manage page
Reduce bottom margin of gift container for better spacing and add a 
PrimaryLinkButton on the manage gift page that links to the gift purchase 
route and opens in a new tab. Refactor PrimaryLinkButton to accept and 
pass the shouldOpenInNewTab argument to support this new behavior. These 
changes improve user flow by making it easier to purchase additional gifts.